Extract main application starter into separate module
